Crate mozjs_unicode_bidi_ffi

Structs

LevelRun

LevelRun type to be returned to C++. 32-bit indexes (rather than usize) are sufficient here because Gecko works with 32-bit indexes when collecting the text buffer for a paragraph.

UnicodeBidi

Bidi object to be exposed to Gecko via FFI.

Functions

bidi_count_runs

Get the number of runs present.

bidi_destroy

Destroy the Bidi object.

bidi_get_base_direction

Get the base direction for the given text, returning 1 for LTR, -1 for RTL, and 0 for neutral. If first_paragraph is true, only the first paragraph will be considered; if false, subsequent paragraphs may be considered until a non-neutral character is found.

bidi_get_direction

Get the paragraph direction: LTR=1, RTL=-1, mixed=0.

bidi_get_length

Get the length of the text covered by the Bidi object.

bidi_get_levels

Get a pointer to the Levels array. The resulting pointer is valid only as long as the UnicodeBidi object exists!

bidi_get_paragraph_level

Get the paragraph level.

bidi_get_visual_run

Get the extent of the run at the given index in the visual runs array. This would panic!() if run_index is out of range (see bidi_count_runs), or if the run’s start or length exceeds u32::MAX (which cannot happen because Gecko can’t create such a huge text buffer).

bidi_new

Create a new UnicodeBidi object for the given text. NOTE that the text buffer must remain valid for the lifetime of this object!

bidi_reorder_visual

Return index map showing the result of reordering using the given levels array. (This is a generic helper that does not use a UnicodeBidi object, it just takes an arbitrary array of levels.)