mozjs_unicode_bidi_ffi

Function bidi_reorder_visual

#[unsafe(no_mangle)]
pub extern "C" fn bidi_reorder_visual(
    levels: *const u8,
    length: usize,
    index_map: *mut i32,
)

Return index map showing the result of reordering using the given levels array. (This is a generic helper that does not use a UnicodeBidi object, it just takes an arbitrary array of levels.)